requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AI/ML Dependencies
google-generativeai>=0.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Web Framework
jinja2>=3.1.2
python-multipart>=0.0.6

# Standard Library (already included in Python)
pathlib
typing
logging


langchain-text-splitters 
sentence-transformers
faiss-cpu 
langchain-google-genai
python-dotenv
PyMuPDF
langchain-anthropic